- Login and logout done - User permissions done

+++ base/views.py
@@ -2,6 +2,8 @@ from django.shortcuts import render, redirect
 from .forms import RoomForm
 from .models import Room, Topic
 from .utils import searchRooms
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 
 
@@ -11,6 +13,7 @@ from .utils import searchRooms
 def home(request):
     topics = Topic.objects.all()
 
+    #function from utils file for search functionality
     query, room = searchRooms(request)
 
     room_count = room.count()
@@ -22,7 +25,7 @@ def home(request):
     }
     return render(request, 'home.html', context)
 
-
+@login_required(login_url='login-user')
 def createRoom(request):
     form = RoomForm()
 
@@ -37,10 +40,14 @@ def createRoom(request):
     }
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login-user')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        messages.error(request, 'You do not have rights to perform this operation')
+        return redirect('home')
 
     if request.method == 'POST':
         form = room = RoomForm(request.POST, instance = room)
@@ -53,9 +60,13 @@ def updateRoom(request,pk):
     }
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login-user')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        messages.error(request, 'You do not have rights to perform this operation')
+        return redirect('home')
 
     if request.method == 'POST':
         room.delete()
